3.46.5

+++ packages/app/src/cli/services/extensions/bundle.ts
@@ -1,23 +1,19 @@
-import {ExtensionBuildOptions} from '../build/extension.js'
-import {ExtensionInstance} from '../../models/extensions/extension-instance.js'
-import {themeExtensionFiles} from '../../utilities/extensions/theme.js'
-import {environmentVariableNames} from '../../constants.js'
+import {buildThemeExtensions, ThemeExtensionBuildOptions} from '../build/extension.js'
 import {context as esContext, BuildResult, formatMessagesSync} from 'esbuild'
 import {AbortSignal} from '@shopify/cli-kit/node/abort'
-import {copyFile} from '@shopify/cli-kit/node/fs'
+import {copyFile, glob} from '@shopify/cli-kit/node/fs'
 import {joinPath, relativePath} from '@shopify/cli-kit/node/path'
-import {outputDebug} from '@shopify/cli-kit/node/output'
-import {isTruthy} from '@shopify/cli-kit/node/context/utilities'
+import {useThemebundling} from '@shopify/cli-kit/node/context/local'
 import {Writable} from 'stream'
 import {createRequire} from 'module'
-import type {StdinOptions, build as esBuild, Plugin} from 'esbuild'
+import type {StdinOptions, build as esBuild} from 'esbuild'
 
 const require = createRequire(import.meta.url)
 
 export interface BundleOptions {
   minify: boolean
   env: {[variable: string]: string}
-  outputPath: string
+  outputBundlePath: string
   stdin: StdinOptions
   stdout: Writable
   stderr: Writable
@@ -49,10 +45,9 @@ export interface BundleOptions {
 /**
  * Invokes ESBuild with the given options to bundle an extension.
  * @param options - ESBuild options
- * @param processEnv - Environment variables for the running process (not those from .env)
  */
-export async function bundleExtension(options: BundleOptions, processEnv = process.env) {
-  const esbuildOptions = getESBuildOptions(options, processEnv)
+export async function bundleExtension(options: BundleOptions) {
+  const esbuildOptions = getESBuildOptions(options)
   const context = await esContext(esbuildOptions)
   if (options.watch) {
     await context.watch()
@@ -69,20 +64,29 @@ export async function bundleExtension(options: BundleOptions, processEnv = proce
   }
 }
 
-export async function bundleThemeExtension(
-  extension: ExtensionInstance,
-  options: ExtensionBuildOptions,
-): Promise<void> {
-  options.stdout.write(`Bundling theme extension ${extension.localIdentifier}...`)
-  const files = await themeExtensionFiles(extension)
+export async function bundleThemeExtensions(options: ThemeExtensionBuildOptions): Promise<void> {
+  if (options.extensions.length === 0) return
 
-  await Promise.all(
-    files.map(function (filepath) {
-      const relativePathName = relativePath(extension.directory, filepath)
-      const outputFile = joinPath(extension.outputPath, relativePathName)
-      return copyFile(filepath, outputFile)
-    }),
-  )
+  await buildThemeExtensions(options)
+
+  if (useThemebundling()) {
+    await Promise.all(
+      options.extensions.map(async (extension) => {
+        options.stdout.write(`Bundling theme extension ${extension.localIdentifier}...`)
+        const files = await glob(joinPath(extension.directory, '/**/*'))
+
+        await Promise.all(
+          files.map(function (filepath) {
+            if (!(filepath.includes('.gitkeep') || filepath.includes('.toml') || filepath.includes('.DS_Store'))) {
+              const relativePathName = relativePath(extension.directory, filepath)
+              const outputFile = joinPath(extension.outputBundlePath, relativePathName)
+              return copyFile(filepath, outputFile)
+            }
+          }),
+        )
+      }),
+    )
+  }
 }
 
 function onResult(result: Awaited<ReturnType<typeof esBuild>> | null, options: BundleOptions) {
@@ -102,7 +106,7 @@ function onResult(result: Awaited<ReturnType<typeof esBuild>> | null, options: B
   }
 }
 
-function getESBuildOptions(options: BundleOptions, processEnv = process.env): Parameters<typeof esContext>[0] {
+function getESBuildOptions(options: BundleOptions): Parameters<typeof esContext>[0] {
   const env: {[variable: string]: string} = options.env
   const define = Object.keys(env || {}).reduce(
     (acc, key) => ({
@@ -112,7 +116,7 @@ function getESBuildOptions(options: BundleOptions, processEnv = process.env): Pa
     {'process.env.NODE_ENV': JSON.stringify(options.environment)},
   )
   const esbuildOptions: Parameters<typeof esContext>[0] = {
-    outfile: options.outputPath,
+    outfile: options.outputBundlePath,
     stdin: options.stdin,
     bundle: true,
     define,
@@ -123,7 +127,7 @@ function getESBuildOptions(options: BundleOptions, processEnv = process.env): Pa
     },
     legalComments: 'none',
     minify: options.minify,
-    plugins: getPlugins(options.stdin.resolveDir, processEnv),
+    plugins: getPlugins(),
     target: 'es6',
     resolveExtensions: ['.tsx', '.ts', '.js', '.json', '.esnext', '.mjs', '.ejs'],
   }
@@ -148,7 +152,7 @@ type ESBuildPlugins = Parameters<typeof esContext>[0]['plugins']
  * It returns the plugins that should be used with ESBuild.
  * @returns List of plugins.
  */
-function getPlugins(resolveDir: string | undefined, processEnv = process.env): ESBuildPlugins {
+function getPlugins(): ESBuildPlugins {
   const plugins = []
 
   if (isGraphqlPackageAvailable()) {
@@ -156,37 +160,7 @@ function getPlugins(resolveDir: string | undefined, processEnv = process.env): E
     plugins.push(graphqlLoader())
   }
 
-  const skipReactDeduplication = isTruthy(processEnv[environmentVariableNames.skipEsbuildReactDedeuplication])
-  if (resolveDir && !skipReactDeduplication) {
-    let resolvedReactPath: string | undefined
-    try {
-      resolvedReactPath = require.resolve('react', {paths: [resolveDir]})
-      // eslint-disable-next-line no-catch-all/no-catch-all
-    } catch {
-      // If weren't able to find React, that's fine. It might not be used.
-      outputDebug(`Unable to load React in ${resolveDir}, skipping React de-duplication`)
-    }
-
-    if (resolvedReactPath) {
-      outputDebug(`Deduplicating React dependency for ${resolveDir}, using ${resolvedReactPath}`)
-      plugins.push(deduplicateReactPlugin(resolvedReactPath))
-    }
-  }
-
   return plugins
-}
-
-function deduplicateReactPlugin(resolvedReactPath: string): Plugin {
-  return {
-    name: 'shopify:deduplicate-react',
-    setup({onResolve}) {
-      onResolve({filter: /^react$/}, (args) => {
-        return {
-          path: resolvedReactPath,
-        }
-      })
-    },
-  }
 }
 
 /**
